Updated event with sonar sound and latest team list
new sonar sound (Joe found, not necessarily official), and reduced team list to 30 officially invited teams.
If teams (like 537), are bringing 2 bots, both must be programmed for their Team# and only 1 of their bots can be on the field at a time. They will only be scheduled for the same number of matches as any other team, so they can either trade off in guaranteed match slots or run the fill-in line